Mistral action examples for deployment
This patch includes custom Mistral actions which are loaded
by the mistral.action entry_point in setup.cfg. The workflows
directory contains an example of how to use these custom
actions in workflows for overcloud deployment.

See env.yaml for an example of the required Mistral environment
format for use with these actions/workflows.

Change-Id: I40738186eca26144b83354fd480da5b6c2f82d2a